requirements.txt
yt-dlp
flask
supabase
python-dotenv
gunicorn

